Improved efficiency in PSOStructures.cpp
Removed a redundent loop in the UpdateSwarmMovement()
Added an extra method named evaluateParticle which helps
in the updateSwarmMovement. With this method implemented, we
don't have to loop twice through the swarm to (a) update their
positions and velocities and (b) evalute their new positions.
The global best fitness as well as the particle best fitness is handled
differently now. When the particles are initialized their respective
fitness gets set to -1, thus when setFitness is called we check if
the fitness is either equal to -1 or smaller than our current fitness.
With the global best, in the contructor of the ConsolePSO and GraphicalPSO
we initialize the Swarm, then we evaulate the Swarm. After the swarm has been
evaluated we set the global swarm best to the first particle in the swarm.
Also for future reference, the reason we evaluate the whole swarm after initialization
and not just the first particle is because the movement of a particle relies on that
it already has a personal best.