Made changes to the hud display.
The HUD now displays the Swarm Best postion as well as the best position
of the selected particle. To make the presentation of the positions fit
within the HUD, the precision of the ostream needed to be changed to 3.
This was accomplished by including the <iomanip> header in gfxmath.
Also added the display of what function is being optimized. This was
accomplished by changing the FunctionDef method to a name() medthod
in DeJongFunctions.cpp & DeJongFunctions.h.